re-structure files and specs to follwo pollsRouter to stay consistent and to have cleaner code
